fix: correct return types in CSI and skill search functions
- Change relevance return type from REAL to NUMERIC in csi.search_masterformat
- Change relevance return type from REAL to INT in public.search_all_skills
- Add verification script to test schema integrity

+++ packages/supabase/migrations/086_create_csi_schema.sql
@@ -64,7 +64,7 @@ RETURNS TABLE (
   code_display TEXT,
   name TEXT,
   depth SMALLINT,
-  relevance REAL
+  relevance NUMERIC
 ) AS $$
 BEGIN
   RETURN QUERY
